(refactor):ajust dos nomes e dirs para best-practices #127
